+++ packages/nextjs/components/builder-registry/bounties/index.tsx
@@ -1,0 +1,2 @@
+export * from "./BountyCard";
+export * from "./BountyStatusTab";
