+++ packages/nextjs/components/builder-registry/contributions/index.tsx
@@ -1,0 +1,4 @@
+export * from "./ContributionCard";
+export * from "./ContributionLikeButton";
+export * from "./ContributionList";
+export * from "./ContributionTable";
