+++ packages/nextjs/components/builder-registry/contributor/index.tsx
@@ -1,3 +1,5 @@
 export * from "./ContributorDetails";
 export * from "./ContributorSkills";
 export * from "./ContributorHeader";
+export * from "./ContributorTable";
+export * from "./ContributorsStats";
